+++ src/main/java/com/berd/dev/services/CategorieDepenseService.java
@@ -1,6 +1,7 @@
 package com.berd.dev.services;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -8,7 +9,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import com.berd.dev.dtos.CategorieDepenseDto;
+import com.berd.dev.mappers.CategorieDepenseMapper;
 import com.berd.dev.models.CategorieDepense;
+import com.berd.dev.models.CategorieDepenseDetail;
+import com.berd.dev.models.User;
 import com.berd.dev.repositories.CategorieDepenseRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -17,6 +22,12 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class CategorieDepenseService {
     private final CategorieDepenseRepository categorieDepenseRepository;
+
+    private final SecurityService securityService;
+
+    public List<CategorieDepenseDto> getAllDto() {
+        return CategorieDepenseMapper.toDto(getAll());
+    }
 
     public List<CategorieDepense> getAll() {
         return categorieDepenseRepository.findAll();
@@ -33,29 +44,65 @@ public class CategorieDepenseService {
     }
 
     public CategorieDepense save(CategorieDepense categorieDepense) {
+        User user = securityService.getAuthenticatedUser();
+        categorieDepense.setUtilisateur(user);
         return categorieDepenseRepository.save(categorieDepense);
     }
 
-    public void deleteById(Integer id) {
+    public void deleteById(Integer id) throws Exception {
+        User user = securityService.getAuthenticatedUser();
+        if (id == null)
+            throw new Exception("Id ne doit pas etre null pendant suppression categorie depense");
+        CategorieDepense cd = categorieDepenseRepository.findById(id).orElse(null);
+        if (cd == null)
+            throw new Exception("Categorie depense non trouvée");
+        if (cd.getUtilisateur() == null || cd.getUtilisateur() != null
+                && !cd.getUtilisateur().getIdUtilisateur().equals(user.getIdUtilisateur()))
+            throw new Exception("Vous ne pouvez supprimez que votre propre categorie depense");
         categorieDepenseRepository.deleteById(id);
     }
 
     public List<CategorieDepense> getByCriteria(String libelle) {
-        System.out.println("libelle: " + libelle);
-        System.out.println(categorieDepenseRepository.findByCriteria(libelle));
-        return categorieDepenseRepository.findByCriteria(libelle);
+        User user = securityService.getAuthenticatedUser();
+        // System.out.println("libelle: " + libelle);
+        // System.out.println(categorieDepenseRepository.findByCriteria(libelle));
+        return categorieDepenseRepository.findByCriteria(libelle, user.getIdUtilisateur());
     }
 
-    public Page<CategorieDepense> getFilteredCategories(String search, String type, int page, int size) {
+    public Page<CategorieDepense> getFilteredCategories(String search, String type, int page, int size, boolean all) {
         Pageable pageable = PageRequest.of(page, size, Sort.by("libelle").ascending());
+        User user = securityService.getAuthenticatedUser();
 
         if ((search != null && !search.trim().isEmpty()) || (type != null && !type.trim().isEmpty())) {
-            return categorieDepenseRepository.findByFilters(
-                    search != null ? search.trim() : "",
-                    type != null ? type.trim() : "",
-                    pageable);
-        }
+            if (all) {
+                return categorieDepenseRepository.findByFiltersAll(
+                        search != null ? search.trim() : "",
+                        type != null ? type.trim() : "",
+                        user.getIdUtilisateur(),
+                        pageable);
+            } else {
+                return categorieDepenseRepository.findByFilters(
+                        search != null ? search.trim() : "",
+                        type != null ? type.trim() : "",
+                        user.getIdUtilisateur(),
+                        pageable);
+            }
 
-        return categorieDepenseRepository.findAll(pageable);
+        }
+         if (all) {
+                 return categorieDepenseRepository.findByFiltersAll(
+                "", // search vide
+                "", // type vide
+                user.getIdUtilisateur(),
+                pageable);
+            } else {
+                return categorieDepenseRepository.findByFilters(
+                "", // search vide
+                "", // type vide
+                user.getIdUtilisateur(),
+                pageable);
+            }
+
+       
     }
 }
